% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCenterOfPeaks}
\alias{getCenterOfPeaks}
\title{Return center of the given granges files}
\usage{
getCenterOfPeaks(gr)
}
\arguments{
\item{gr}{granges coordinate}
}
\value{
granges format of the center
}
\description{
Get a granges and find the center of it
}
