% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewiseAssociations.R
\name{getNameFromCoordinates}
\alias{getNameFromCoordinates}
\title{Get names of genes or peaks related to a query coordinates}
\usage{
getNameFromCoordinates(names, coordinates, queryCoordinates)
}
\arguments{
\item{names}{Names of genes or peaks}

\item{coordinates}{Coordinates of genes or peaks in granges format}

\item{queryCoordinates}{Coordinates of the query regions in granges format}
}
\value{
Names of genes or peaks in queried reiongs
}
\description{
Get names and coordinates of genes or peaks. It also get the

coordinates of query regions and returns the realted genes or

peak names.
}
