% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{granges2String}
\alias{granges2String}
\title{Conver granes to strings of cooridnates}
\usage{
granges2String(gr)
}
\arguments{
\item{gr}{granges coordinates}
}
\value{
string of coordinates
}
\description{
Get genomic coordinates granges and convert them to strings
}
