% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewiseAssociations.R
\name{removeReserveCharacter}
\alias{removeReserveCharacter}
\title{Remove reserved characters from a string}
\usage{
removeReserveCharacter(name)
}
\arguments{
\item{name}{A string of characters}
}
\value{
A string without reserved characters
}
\description{
Remove reserved characters (such as *, +, -, etc) from a string
}
\examples{
removeReserveCharacter("A&\%B^f6")
}
