% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{site2GeneDistance}
\alias{site2GeneDistance}
\title{Return the distance between paired peaks and genes}
\usage{
site2GeneDistance(geneCoordinates, peakCoordinates)
}
\arguments{
\item{geneCoordinates}{granges coordinates of genes}

\item{peakCoordinates}{granges coordinates of peaks}
}
\value{
the respective distances of paired genes and peaks
}
\description{
Get a granges of genes and peaks and return their distances
}
