% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string2Granges}
\alias{string2Granges}
\title{Convert strings to granges of coordinates}
\usage{
string2Granges(strCoordinates)
}
\arguments{
\item{strCoordinates}{string of coordinates}
}
\value{
Genomic coordinates in granges format
}
\description{
Get genomic coordinates as strings and convert them to granges
}
\examples{
string2Granges(c("chr1:1112-1231", "ch2:3131-3221"))

}
