% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string2Granges}
\alias{string2Granges}
\title{Conver strings to granges of cooridnates}
\usage{
string2Granges(strCoordinates)
}
\arguments{
\item{strCoordinates}{string of coordinates}
}
\value{
Genomic coordinates in granges format
}
\description{
Get genomic coordinates as trings and convert them to grangess
}
